I want to be transparent about a few things regarding this contribution:

C Expertise: I am not a C developer by trade, so I’ve done my best to follow the project's style and logic, but I might have missed some language-specific nuances.

AI Assistance: I used AI tools to help draft certain parts of the logic and ensure the syntax was correct.

Context: This is a hobby project done during my free time.

Feedback Welcome

Since I’m learning as I go, I am very open to constructive feedback. If you see ways to make the code more "idiomatic" C or if something doesn't align with your standards, please let me know!

I would also love it if anyone could test these changes in their own environment. Detailed bug reports or suggestions for improvement are more than welcome.


I hope one day push this to main repo.


#fix https://github.com/eclipse-paho/paho.mqtt.embedded-c/issues/274

+++ MQTTPacket/src/V3/MQTTFormat.h
@@ -17,8 +17,8 @@
 #if !defined(MQTTFORMAT_H)
 #define MQTTFORMAT_H
 
-#include "StackTrace.h"
-#include "MQTTPacket.h"
+#include "../StackTrace.h"
+#include "../MQTTPacket.h"
 
 const char* MQTTPacket_getName(unsigned short packetid);
 int MQTTStringFormat_connect(char* strbuf, int strbuflen, MQTTPacket_connectData* data);
@@ -31,7 +31,7 @@ int MQTTStringFormat_subscribe(char* strbuf, int strbuflen, unsigned char dup, u
 int MQTTStringFormat_suback(char* strbuf, int strbuflen, unsigned short packetid, int count, unsigned char* grantedQoSs);
 int MQTTStringFormat_unsubscribe(char* strbuf, int strbuflen, unsigned char dup, unsigned short packetid,
 		int count, MQTTString topicFilters[]);
-char* MQTTFormat_toClientString(char* strbuf, int strbuflen, unsigned char* buf, int32_t buflen);
-char* MQTTFormat_toServerString(char* strbuf, int strbuflen, unsigned char* buf, int32_t buflen);
+char* MQTTFormat_toClientString(char* strbuf, int strbuflen, unsigned char* buf, size_t buflen);
+char* MQTTFormat_toServerString(char* strbuf, int strbuflen, unsigned char* buf, size_t buflen);
 
 #endif
